Updated searchLogins: set realm to null if caught
LoginHelper.jsm > doLoginsMatch() checks the httpRealm property of the submitted form login and the passifox login. All tests had formLogin.httpRealm as null, but passifox login.httpRealm was an empty string, as the realm was being caught and set in searchLogins.

Updated searchLogins to have realm set to null if caught.